backend/requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-CORS==4.0.0
Werkzeug==2.3.0
mysql-connector-python==8.0.33